Juntar las definiciones en un archivo
