+++ src/asignacion_aulica/gestor_de_datos/__init__.py
@@ -1,5 +1,4 @@
-from asignacion_aulica.gestor_de_datos.enums import Día, PeriodoDeClases
+from asignacion_aulica.gestor_de_datos.entidades import Aula, Edificio, Carrera, Clase
+from asignacion_aulica.lógica_de_asignación import AsignaciónImposibleException
 from asignacion_aulica.gestor_de_datos.gestor import GestorDeDatos
-from asignacion_aulica.gestor_de_datos.edificio import Edificio
-from asignacion_aulica.gestor_de_datos.clase import Clase
-from asignacion_aulica.gestor_de_datos.aula import Aula
+from asignacion_aulica.gestor_de_datos.día import Día
