--- src/asignacion_aulica/gestor_de_datos/carrera.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional
-
-@dataclass
-class Carrera:
-    nombre: str
-    edificio_preferido: Optional[str]
