+++ src/asignacion_aulica/gestor_de_datos/gestor.py
@@ -1,4 +1,6 @@
+from asignacion_aulica.gestor_de_datos.entidades import Aula, Edificio, Carrera, Clase
 from asignacion_aulica.lógica_de_asignación import AsignaciónImposibleException
+from asignacion_aulica.gestor_de_datos.enums import Día, PeriodoDeClases
 
 class GestorDeDatos:
     '''
